requirements.txt
requests
python-dotenv
aiohttp
tqdm
simple-term-menu
